LangChain/requirements.txt
ipykernel
numpy
pandas
langchain
langchain-community
python-dotenv
langchain-groq
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-openai
langchain-ollama
langchain-huggingface
sentence-transformers
faiss-cpu
langchain_chroma
pysqlite3-binary
chromadb
streamlit
fastapi
langserve
uvicorn
requests